Fix computer disconnects due to too early VM stop
check() method in retention strategy may be called after VM start but
before launching slave on it. Plain isIdle() check returns true -- and
we started disconnecting and removing node, _before_ initial connecting!
So computer start fails and job is not executed.

To fix this lets introduce idle timeout which should be passed before we
start computer termination. As it is done in retention strategies of
other plugins. 2 minutes value is "taken from the ceiling".