Don't abort if unable to get VM IP address
Previously we aborted VM computer launch if were unable to get its IP
address via `prlctl list -f`. This could happen because Parallels Tools
are not installed in that VM or just because of a bug in Parallels
Desktop. But we may use a fallback way: just try to use address of
hostname specified in a form of VM configuration.

Bug #JENKINS-41431